requirements.txt
requests==2.28.2
tweepy==4.12.1
python-dotenv==0.21.0
